reverted match constants and updated based on review
