+++ src/constants/index.ts
@@ -1,3 +1,3 @@
 export * from "./fake-data/index.js";
 export * from "./main.js";
-export * from "./characters.jsx";
+export * from "./characters.js";
