+++ src/constants/main.ts
@@ -28,11 +28,3 @@ export const POINTS_ACHIEVEMENTS: Record<AchievementId, number> = {
   "201": 5,
   "202": 5,
 };
-
-//TODO: replace all match stages with these constants
-
-/**** MATCH ****/
-export const MATCH_HOST = "host" as const;
-export const VOTING = "voting" as const;
-export const RESULTS = "results" as const;
-export const CHAT = "chat" as const;
