+++ src/pages/match.tsx
@@ -11,7 +11,6 @@ import { Results } from "~/components/results/index.js";
 import { ErrorView } from "~/components/error-view/index.jsx";
 import { PlayerLocal } from "~/components/players/player-local/index.js";
 import { PlayersOthers } from "~/components/players/players-others/index.js";
-import { CHAT, RESULTS } from "~/constants/index.js";
 
 const Match: FC = () => {
   const { query } = useRouter();
@@ -70,7 +69,7 @@ const MatchInternal: FC<Props> = ({ roomId, session }) => {
     await vote.mutateAsync({ selectedUserIds, roomId });
   };
 
-  const splashVariant = room.stage === RESULTS ? undefined : room.stage;
+  const splashVariant = room.stage === "results" ? undefined : room.stage;
 
   return (
     <Layout splashScreenVariant={splashVariant} localPlayer={localPlayer}>
@@ -80,7 +79,7 @@ const MatchInternal: FC<Props> = ({ roomId, session }) => {
         isVoteEnabled={isVoteEnabled}
         onVote={handleVote}
       />
-      {room.stage === RESULTS && (
+      {room.stage === "results" && (
         <Results
           gainedScore={localPlayer.score}
           totalBots={totalBots}
@@ -89,8 +88,8 @@ const MatchInternal: FC<Props> = ({ roomId, session }) => {
         />
       )}
 
-      {room.stage !== RESULTS && <Chat roomId={roomId} room={room} />}
-      {room.stage === CHAT && <PlayerLocal localPlayer={localPlayer} />}
+      {room.stage !== "results" && <Chat roomId={roomId} room={room} />}
+      {room.stage === "chat" && <PlayerLocal localPlayer={localPlayer} />}
     </Layout>
   );
 };
