+++ src/styles/index.ts
@@ -1,1 +1,4 @@
 export * from "./provider.jsx";
+export * from "./palette.js";
+export * from "./theme.js";
+export * from "./typography.js";
